fix(github-sync): revert ParentMap key to issue number
gh issue list --json does not support databaseId — the field is absent
from the GraphQL schema exposed by that command. Using it silently zeroed
every child key, producing an empty ParentMap and no parent links.

Reverts the child key back to issue number (map[int]int64) and the
--json flag back to "number,parent".

+++ cmd/kata/daemon_cmd_test.go
@@ -1060,7 +1060,7 @@ func (f *daemonGitHubSyncFetcher) Comments(_ context.Context, _ githubsync.Bindi
 	return append([]githubsync.Comment(nil), f.comments[issueNumber]...), nil
 }
 
-func (f *daemonGitHubSyncFetcher) ParentMap(_ context.Context, _ githubsync.Binding) (map[int64]int64, error) {
+func (f *daemonGitHubSyncFetcher) ParentMap(_ context.Context, _ githubsync.Binding) (map[int]int64, error) {
 	return nil, nil
 }
